% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_document.R
\name{su_document}
\alias{su_document}
\title{Convert to an Strategy Unit Word document}
\usage{
su_document(
  use_numbered_headings = FALSE,
  toc = FALSE,
  toc_depth = 1,
  fig_caption = TRUE,
  df_print = "kable",
  highlight = "default",
  keep_md = FALSE,
  md_extensions = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{use_numbered_headings}{should we use numbered headings? Defaults to
false}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{df_print}{Method to be used for printing data frames. Valid values
include "default", "kable", "tibble", and "paged". The "default" method
uses a corresponding S3 method of \code{print}, typically
\code{print.data.frame}. The "kable" method uses the
\code{\link[knitr:kable]{knitr::kable}} function. The "tibble" method uses
the \pkg{tibble} package to print a summary of the data frame. The "paged"
method creates a paginated HTML table (note that this method is only valid
for formats that produce HTML). In addition to the named methods you can
also pass an arbitrary function to be used for printing data frames. You
can disable the \code{df_print} behavior entirely by setting the option
\code{rmarkdown.df_print} to \code{FALSE}.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a Strategy Unit branded MS Word
document.
}
\details{
See the \href{https://rmarkdown.rstudio.com/word_document_format.html}{online
documentation} for additional details on using the \code{word_document}
format.
}
\examples{
\dontrun{
library(rmarkdown)

# simple invocation
render("input.Rmd", su_document())
}
}
