% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_theme.R
\name{su_theme}
\alias{su_theme}
\title{Creates the Strategy Unit Theme for ggplot}
\usage{
su_theme(base_family = "Segoe UI")
}
\arguments{
\item{base_family}{The font to use in this theme}
}
\description{
Creates a base theme for ggplot objects. By default this uses a font that
requires the \link[extrafont]{extrafont} package to be installed.
}
\examples{
\dontrun{
library(ggplot2)
library(StrategyUnitTheme)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  su_theme()
}
}
