% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_theme_cols.R
\docType{data}
\name{su_theme_colours}
\alias{su_theme_colours}
\title{Strategy Unit Theme Colours Vector}
\format{An object of class \code{character} of length 24.}
\usage{
su_theme_colours
}
\description{
The colours that are used in the Strategy Unit Theme. Non-exported (used by
other functions e.g. su_theme_cols)
}
\keyword{datasets}
