% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_theme_gradient_colour_scales.R
\name{su_theme_gradient_colour_scales}
\alias{su_theme_gradient_colour_scales}
\alias{scale_colour_su}
\alias{scale_color_su}
\alias{scale_fill_su}
\title{Strategy Unit Theme Gradient colour Scale}
\usage{
scale_colour_su(
  palette = c("main", "oranges", "slates", "mixed", "oj_coal", "oj_red"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_color_su(
  palette = c("main", "oranges", "slates", "mixed", "oj_coal", "oj_red"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_fill_su(
  palette = c("main", "oranges", "slates", "mixed", "oj_coal", "oj_red"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The name of palette to use}

\item{discrete}{Optional: boolean to indicate that this scale is a discrete
scale. Defaults to TRUE}

\item{reverse}{Optional: boolean to reverse the direction of the scale.
Defaults to FALSE}

\item{...}{additional arguments passed to the ggplot functions}
}
\value{
a scale object to be used with a ggplot object
}
\description{
Generates a two colour gradient (low-high)
}
