% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_theme_pal.R
\name{su_theme_pal}
\alias{su_theme_pal}
\title{Strategy Unit Theme Palettes}
\usage{
su_theme_pal(palette = su_theme_palettes, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The name of palette to use}

\item{reverse}{Optional: boolean value to reverse the palette}

\item{...}{arguments to pass to colorRamp}
}
\value{
A colorRamp function
}
\description{
This function returns a palette generated by colorRampPalette from one of the
Strategy Unit palettes
}
\examples{
su_theme_pal()
# or, specify a paltte
su_theme_pal("oranges")
# a reverse palette
su_theme_pal("slates", TRUE)
}
