% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_theme.R
\name{su_theme}
\alias{su_theme}
\title{Creates the Strategy Unit Theme for ggplot}
\usage{
su_theme()
}
\description{
Creates a base theme for ggplot objects. Make sure that you have installed
the \link[extrafont]{extrafont} package and
\href{https://blog.revolutionanalytics.com/2012/09/how-to-use-your-favorite-fonts-in-r-charts.html}{set it up}.
}
